requirements.txt
# Web会議録画 ナレッジレポート自動生成システム
# Python依存パッケージ

# 注意: 以下の外部ツールはpipでインストールできません
# 別途インストールが必要です:
# - ffmpeg: 画像抽出用
# - pandoc: ドキュメント変換用

# 画像処理用（detect_scene_changes.py で使用）
Pillow>=10.0.0

# テスト用（オプション）
# pytest>=7.0.0

# コードフォーマット用（オプション）
# black>=23.0.0
# flake8>=6.0.0